requirements.txt
# =============================================================================
# ScaleCast Dependencies
# =============================================================================
# MLOps Pipeline for Demand Forecasting
# Pin all versions for reproducibility

# =============================================================================
# Workflow Orchestration
# =============================================================================
apache-airflow==2.7.3

# =============================================================================
# Data Processing
# =============================================================================
pandas==2.1.3
numpy>=1.24.0,<2.0.0

# =============================================================================
# Machine Learning
# =============================================================================
scikit-learn==1.3.2
mlflow==2.9.1

# =============================================================================
# API Framework
# =============================================================================
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# =============================================================================
# Data Validation
# =============================================================================
great-expectations==0.18.3

# =============================================================================
# Data Version Control
# =============================================================================
dvc[s3]==3.30.1

# =============================================================================
# AWS Integration
# =============================================================================
boto3==1.33.6

# =============================================================================
# Database
# =============================================================================
psycopg2-binary==2.9.9
sqlalchemy>=1.4.0,<3.0.0

# =============================================================================
# Configuration & Utilities
# =============================================================================
python-dotenv==1.0.0
pyyaml>=6.0
httpx==0.25.2

# =============================================================================
# Security & Cryptography
# =============================================================================
cryptography>=41.0.0

# =============================================================================
# Testing
# =============================================================================
pytest==7.4.3
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
